fix(ci): split oversized i18n-touched components for max-lines
Extract focused sibling modules from CreatePullRequestDialog,
SparsePresetSettingsSection, and BrowserToolbarMenu so oxlint
max-lines passes without disabling the rule.
