+++ src/renderer/src/components/browser-pane/BrowserToolbarMenu.tsx
@@ -1,39 +1,15 @@
 import { useLayoutEffect, useState } from 'react'
-import { Check, Ellipsis, Import, Monitor, Plus, Settings } from 'lucide-react'
 import { toast } from 'sonner'
-import { Button } from '@/components/ui/button'
-import { Input } from '@/components/ui/input'
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle
-} from '@/components/ui/dialog'
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuPortal,
-  DropdownMenuRadioGroup,
-  DropdownMenuRadioItem,
-  DropdownMenuSeparator,
-  DropdownMenuSub,
-  DropdownMenuSubContent,
-  DropdownMenuSubTrigger,
-  DropdownMenuTrigger
-} from '@/components/ui/dropdown-menu'
 import { useAppStore } from '@/store'
 import { useMountedRef } from '@/hooks/useMountedRef'
 import { shouldShowBrowserImportHint } from './browser-import-hint-visibility'
-import { BROWSER_FAMILY_LABELS } from '../../../../shared/constants'
 import type { BrowserViewportPresetId } from '../../../../shared/types'
 import {
-  BROWSER_VIEWPORT_PRESETS,
   browserViewportPresetToOverride,
   getBrowserViewportPreset
 } from '../../../../shared/browser-viewport-presets'
+import { BrowserToolbarMenuDropdown } from './browser-toolbar-menu-dropdown'
+import { BrowserToolbarProfileDialogs } from './browser-toolbar-profile-dialogs'
 import { translate } from '@/i18n/i18n'
 
 type BrowserToolbarMenuProps = {
@@ -226,280 +202,39 @@ export function BrowserToolbarMenu({
 
   return (
     <>
-      <DropdownMenu modal={false} open={menuOpen} onOpenChange={handleMenuOpenChange}>
-        <DropdownMenuTrigger asChild>
-          <Button
-            size="icon"
-            variant="ghost"
-            className="h-8 w-8"
-            title={translate(
-              'auto.components.browser.pane.BrowserToolbarMenu.7b838540c7',
-              'Browser menu'
-            )}
-          >
-            <Ellipsis className="size-4" />
-          </Button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent align="end" className="w-56">
-          {allProfiles.map((profile) => {
-            const isSelectedProfile = profile.id === effectiveProfileId
-            return (
-              <DropdownMenuItem
-                key={profile.id}
-                onSelect={() => handleSwitchProfile(profile.id === 'default' ? null : profile.id)}
-              >
-                <Check
-                  className={`mr-2 size-3.5 shrink-0 ${isSelectedProfile ? 'opacity-100' : 'opacity-0'}`}
-                />
-                <span className="truncate">{profile.label}</span>
-                {profile.source?.browserFamily && (
-                  <span className="ml-auto pl-2 text-[10px] text-muted-foreground">
-                    {BROWSER_FAMILY_LABELS[profile.source.browserFamily] ??
-                      profile.source.browserFamily}
-                  </span>
-                )}
-              </DropdownMenuItem>
-            )
-          })}
+      <BrowserToolbarMenuDropdown
+        menuOpen={menuOpen}
+        onMenuOpenChange={handleMenuOpenChange}
+        allProfiles={allProfiles}
+        effectiveProfileId={effectiveProfileId}
+        onSwitchProfile={handleSwitchProfile}
+        onNewProfile={() => setNewProfileDialogOpen(true)}
+        detectedBrowsers={detectedBrowsers}
+        onFetchDetectedBrowsers={() => void fetchDetectedBrowsers()}
+        browserSessionImportState={browserSessionImportState}
+        onImportFromBrowser={(browserFamily, browserProfile) =>
+          void handleImportFromBrowser(browserFamily, browserProfile)
+        }
+        onImportFromFile={() => void handleImportFromFile()}
+        viewportPresetId={viewportPresetId}
+        onApplyViewportPreset={applyViewportPreset}
+      />
 
-          <DropdownMenuSeparator />
-
-          <DropdownMenuItem onSelect={() => setNewProfileDialogOpen(true)}>
-            <Plus className="mr-2 size-3.5" />
-            {translate(
-              'auto.components.browser.pane.BrowserToolbarMenu.cf7cdc67ef',
-              'New Profile…'
-            )}
-          </DropdownMenuItem>
-
-          <DropdownMenuSeparator />
-
-          <DropdownMenuSub
-            onOpenChange={(open) => {
-              if (open) {
-                // Why: macOS treats other browsers' profile folders as app
-                // data. Only probe them when the user opens the import menu.
-                void fetchDetectedBrowsers()
-              }
-            }}
-          >
-            <DropdownMenuSubTrigger
-              disabled={
-                browserSessionImportState?.profileId === effectiveProfileId &&
-                browserSessionImportState.status === 'importing'
-              }
-              data-contextual-tour-target="browser-import-cookies-control"
-            >
-              <Import className="mr-2 size-3.5" />
-              {translate(
-                'auto.components.browser.pane.BrowserToolbarMenu.2293adf620',
-                'Import Cookies'
-              )}
-            </DropdownMenuSubTrigger>
-            <DropdownMenuPortal>
-              <DropdownMenuSubContent>
-                {detectedBrowsers.map((browser) =>
-                  browser.profiles.length > 1 ? (
-                    <DropdownMenuSub key={browser.family}>
-                      <DropdownMenuSubTrigger>
-                        {translate(
-                          'auto.components.browser.pane.BrowserToolbarMenu.eb280bfb11',
-                          'From'
-                        )}
-                        {browser.label}
-                      </DropdownMenuSubTrigger>
-                      <DropdownMenuPortal>
-                        <DropdownMenuSubContent>
-                          {browser.profiles.map((profile) => (
-                            <DropdownMenuItem
-                              key={profile.directory}
-                              onSelect={() =>
-                                void handleImportFromBrowser(browser.family, profile.directory)
-                              }
-                            >
-                              {profile.name}
-                            </DropdownMenuItem>
-                          ))}
-                        </DropdownMenuSubContent>
-                      </DropdownMenuPortal>
-                    </DropdownMenuSub>
-                  ) : (
-                    <DropdownMenuItem
-                      key={browser.family}
-                      onSelect={() => void handleImportFromBrowser(browser.family)}
-                    >
-                      {translate(
-                        'auto.components.browser.pane.BrowserToolbarMenu.eb280bfb11',
-                        'From'
-                      )}
-                      {browser.label}
-                    </DropdownMenuItem>
-                  )
-                )}
-                {detectedBrowsers.length > 0 && <DropdownMenuSeparator />}
-                <DropdownMenuItem onSelect={() => void handleImportFromFile()}>
-                  {translate(
-                    'auto.components.browser.pane.BrowserToolbarMenu.56f94f4ffa',
-                    'From File…'
-                  )}
-                </DropdownMenuItem>
-              </DropdownMenuSubContent>
-            </DropdownMenuPortal>
-          </DropdownMenuSub>
-
-          <DropdownMenuSeparator />
-
-          <DropdownMenuSub>
-            <DropdownMenuSubTrigger>
-              <Monitor className="mr-2 size-3.5" />
-              {translate(
-                'auto.components.browser.pane.BrowserToolbarMenu.e5d31de1a9',
-                'Viewport Size'
-              )}
-            </DropdownMenuSubTrigger>
-            <DropdownMenuPortal>
-              <DropdownMenuSubContent>
-                {/* Why: Viewport is a "pick one of N" control, so use a radio group
-                    for proper a11y semantics (role="menuitemradio", aria-checked).
-                    The "Default" option represents a null preset (no override),
-                    encoded as the sentinel string 'default' because
-                    DropdownMenuRadioGroup values must be strings. */}
-                <DropdownMenuRadioGroup
-                  value={viewportPresetId ?? 'default'}
-                  onValueChange={(v) =>
-                    applyViewportPreset(v === 'default' ? null : (v as BrowserViewportPresetId))
-                  }
-                >
-                  <DropdownMenuRadioItem value="default">
-                    {translate(
-                      'auto.components.browser.pane.BrowserToolbarMenu.ed8f54509d',
-                      'Default'
-                    )}
-                  </DropdownMenuRadioItem>
-                  <DropdownMenuSeparator />
-                  {BROWSER_VIEWPORT_PRESETS.map((preset) => (
-                    <DropdownMenuRadioItem key={preset.id} value={preset.id}>
-                      <span className="truncate">{preset.label}</span>
-                    </DropdownMenuRadioItem>
-                  ))}
-                </DropdownMenuRadioGroup>
-              </DropdownMenuSubContent>
-            </DropdownMenuPortal>
-          </DropdownMenuSub>
-
-          <DropdownMenuSeparator />
-
-          <DropdownMenuItem
-            onSelect={() => {
-              useAppStore.getState().openSettingsTarget({ pane: 'browser', repoId: null })
-              useAppStore.getState().openSettingsPage()
-            }}
-          >
-            <Settings className="mr-2 size-3.5" />
-            {translate(
-              'auto.components.browser.pane.BrowserToolbarMenu.a771c2b6c8',
-              'Browser Settings…'
-            )}
-          </DropdownMenuItem>
-        </DropdownMenuContent>
-      </DropdownMenu>
-
-      <Dialog
-        open={pendingSwitchProfileId !== undefined}
-        onOpenChange={(open) => {
-          if (!open) {
-            setPendingSwitchProfileId(undefined)
-          }
+      <BrowserToolbarProfileDialogs
+        pendingSwitchProfileId={pendingSwitchProfileId}
+        onPendingSwitchChange={() => setPendingSwitchProfileId(undefined)}
+        onConfirmSwitch={confirmSwitchProfile}
+        newProfileDialogOpen={newProfileDialogOpen}
+        onNewProfileDialogOpenChange={setNewProfileDialogOpen}
+        newProfileName={newProfileName}
+        onNewProfileNameChange={setNewProfileName}
+        isCreatingProfile={isCreatingProfile}
+        onCreateProfile={() => void handleCreateProfile()}
+        onCancelNewProfile={() => {
+          setNewProfileDialogOpen(false)
+          setNewProfileName('')
         }}
-      >
-        <DialogContent className="sm:max-w-sm" showCloseButton={false}>
-          <DialogHeader>
-            <DialogTitle className="text-base">
-              {translate(
-                'auto.components.browser.pane.BrowserToolbarMenu.fe683eb3b4',
-                'Switch Profile'
-              )}
-            </DialogTitle>
-            <DialogDescription className="text-xs">
-              {translate(
-                'auto.components.browser.pane.BrowserToolbarMenu.a38f217b46',
-                'Switching profiles will reload this page. Any unsaved form data will be lost.'
-              )}
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => setPendingSwitchProfileId(undefined)}
-            >
-              {translate('auto.components.browser.pane.BrowserToolbarMenu.429ef481f9', 'Cancel')}
-            </Button>
-            <Button size="sm" onClick={confirmSwitchProfile}>
-              {translate('auto.components.browser.pane.BrowserToolbarMenu.58f2c81542', 'Switch')}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-
-      <Dialog open={newProfileDialogOpen} onOpenChange={setNewProfileDialogOpen}>
-        <DialogContent className="sm:max-w-sm" showCloseButton={false}>
-          <DialogHeader>
-            <DialogTitle className="text-base">
-              {translate(
-                'auto.components.browser.pane.BrowserToolbarMenu.67e9b9fcd6',
-                'New Browser Profile'
-              )}
-            </DialogTitle>
-          </DialogHeader>
-          <form
-            onSubmit={(e) => {
-              e.preventDefault()
-              void handleCreateProfile()
-            }}
-          >
-            <Input
-              value={newProfileName}
-              onChange={(e) => setNewProfileName(e.target.value)}
-              placeholder={translate(
-                'auto.components.browser.pane.BrowserToolbarMenu.64f448fb6e',
-                'Profile name'
-              )}
-              autoFocus
-              maxLength={50}
-              className="mb-4"
-            />
-            <DialogFooter>
-              <Button
-                type="button"
-                variant="outline"
-                size="sm"
-                onClick={() => {
-                  setNewProfileDialogOpen(false)
-                  setNewProfileName('')
-                }}
-              >
-                {translate('auto.components.browser.pane.BrowserToolbarMenu.429ef481f9', 'Cancel')}
-              </Button>
-              <Button
-                type="submit"
-                size="sm"
-                disabled={!newProfileName.trim() || isCreatingProfile}
-              >
-                {isCreatingProfile
-                  ? translate(
-                      'auto.components.browser.pane.BrowserToolbarMenu.bf648471c5',
-                      'Creating…'
-                    )
-                  : translate(
-                      'auto.components.browser.pane.BrowserToolbarMenu.569bce8eb1',
-                      'Create'
-                    )}
-              </Button>
-            </DialogFooter>
-          </form>
-        </DialogContent>
-      </Dialog>
+      />
     </>
   )
 }
