+++ src/renderer/src/components/right-sidebar/CreatePullRequestDialog.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useRef, useState } from 'react'
-import { Check, ChevronsUpDown, Loader2 } from 'lucide-react'
+import { Loader2 } from 'lucide-react'
 import { toast } from 'sonner'
 import { Button } from '@/components/ui/button'
 import {
@@ -10,12 +10,8 @@ import {
   DialogHeader,
   DialogTitle
 } from '@/components/ui/dialog'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
-import { cn } from '@/lib/utils'
 import { useAppStore } from '@/store'
 import type {
-  CreateHostedReviewResult,
   HostedReviewCreationEligibility,
   HostedReviewProvider
 } from '../../../../shared/hosted-review'
@@ -29,6 +25,8 @@ import {
 import { getCommitMessageModelDiscoveryHostKeyForScope } from '../../../../shared/commit-message-host-key'
 import { getRuntimeGitScope } from '@/runtime/runtime-git-client'
 import { CreatePullRequestGenerateButton } from './CreatePullRequestGenerateButton'
+import { CreatePullRequestDialogForm } from './CreatePullRequestDialogForm'
+import { formatCreateError, reviewCopy } from './create-pull-request-review-copy'
 import { translate } from '@/i18n/i18n'
 
 type CreatePullRequestDialogProps = {
@@ -48,42 +46,6 @@ type CreatePullRequestDialogProps = {
     number: number
     url: string
   }) => Promise<void>
-}
-
-function reviewCopy(provider: HostedReviewProvider): {
-  shortLabel: 'PR' | 'MR'
-  reviewLabel: 'pull request' | 'merge request'
-  titleLabel: 'Pull Request' | 'Merge Request'
-  providerName: 'GitHub' | 'GitLab'
-} {
-  return provider === 'gitlab'
-    ? {
-        shortLabel: 'MR',
-        reviewLabel: 'merge request',
-        titleLabel: 'Merge Request',
-        providerName: 'GitLab'
-      }
-    : {
-        shortLabel: 'PR',
-        reviewLabel: 'pull request',
-        titleLabel: 'Pull Request',
-        providerName: 'GitHub'
-      }
-}
-
-function formatCreateError(
-  result: CreateHostedReviewResult,
-  pushed: boolean,
-  shortLabel: 'PR' | 'MR'
-): string {
-  if (result.ok) {
-    return ''
-  }
-  if (pushed) {
-    const prefix = new RegExp(`^Create ${shortLabel} failed:\\s*`, 'i')
-    return `Push succeeded, but ${shortLabel} creation failed: ${result.error.replace(prefix, '')}`
-  }
-  return result.error
 }
 
 export function CreatePullRequestDialog({
@@ -326,148 +288,25 @@ export function CreatePullRequestDialog({
           </DialogDescription>
         </DialogHeader>
 
-        <div className="space-y-4">
-          <div className="space-y-1">
-            <Label>
-              {translate(
-                'auto.components.right.sidebar.CreatePullRequestDialog.6f5f1962b6',
-                'Head branch'
-              )}
-            </Label>
-            <div className="inline-flex max-w-full items-center rounded-full border border-border bg-muted px-2 py-1 text-xs font-medium text-foreground">
-              <span className="truncate">{branch}</span>
-            </div>
-          </div>
-
-          <div className="space-y-2">
-            <div className="space-y-1">
-              <Label htmlFor="create-pr-base">
-                {translate(
-                  'auto.components.right.sidebar.CreatePullRequestDialog.8584ccb43c',
-                  'Base branch'
-                )}
-              </Label>
-              <p className="text-xs text-muted-foreground">
-                {translate(
-                  'auto.components.right.sidebar.CreatePullRequestDialog.0fad57a14c',
-                  'Search remote branches or enter a branch name.'
-                )}
-              </p>
-            </div>
-            <div className="relative">
-              <Input
-                id="create-pr-base"
-                value={baseQuery || base}
-                onChange={(event) => {
-                  setBaseQuery(event.target.value)
-                  setBase(event.target.value)
-                }}
-                placeholder={translate(
-                  'auto.components.right.sidebar.CreatePullRequestDialog.694550a610',
-                  'main'
-                )}
-                aria-invalid={!base.trim()}
-                className="pr-8"
-              />
-              <ChevronsUpDown className="pointer-events-none absolute right-2 top-2.5 size-3.5 text-muted-foreground" />
-            </div>
-            {baseSearchError ? <p className="text-xs text-destructive">{baseSearchError}</p> : null}
-            {baseResults.length > 0 ? (
-              <div className="max-h-36 overflow-auto rounded-md border border-border p-1 scrollbar-sleek">
-                {baseResults.map((ref) => (
-                  <button
-                    key={ref}
-                    type="button"
-                    className={cn(
-                      'flex w-full items-center justify-between rounded-sm px-2 py-1.5 text-left text-xs hover:bg-accent',
-                      stripBaseRef(base) === ref && 'bg-accent text-accent-foreground'
-                    )}
-                    onClick={() => {
-                      setBase(ref)
-                      setBaseQuery('')
-                      setBaseResults([])
-                    }}
-                  >
-                    <span className="truncate">{ref}</span>
-                    {stripBaseRef(base) === ref ? <Check className="size-3" /> : null}
-                  </button>
-                ))}
-              </div>
-            ) : null}
-          </div>
-
-          <div className="space-y-2">
-            <Label htmlFor="create-pr-title">
-              {translate(
-                'auto.components.right.sidebar.CreatePullRequestDialog.68314b4369',
-                'Title'
-              )}
-            </Label>
-            <Input
-              id="create-pr-title"
-              value={title}
-              onChange={(event) => setTitle(event.target.value)}
-              placeholder={translate(
-                'auto.components.right.sidebar.CreatePullRequestDialog.68314b4369',
-                'Title'
-              )}
-              aria-invalid={!title.trim()}
-            />
-          </div>
-
-          <div className="space-y-2">
-            <Label htmlFor="create-pr-body">
-              {translate(
-                'auto.components.right.sidebar.CreatePullRequestDialog.1cd53359db',
-                'Description'
-              )}
-            </Label>
-            <textarea
-              id="create-pr-body"
-              value={body}
-              onChange={(event) => setBody(event.target.value)}
-              rows={6}
-              placeholder={translate(
-                'auto.components.right.sidebar.CreatePullRequestDialog.02b2ce911f',
-                'Description (optional)'
-              )}
-              className="w-full resize-none rounded-md border border-border bg-background px-3 py-2 text-sm text-foreground outline-none placeholder:text-muted-foreground/70 focus-visible:ring-1 focus-visible:ring-ring"
-            />
-            <p className="text-xs text-muted-foreground">
-              {translate(
-                'auto.components.right.sidebar.CreatePullRequestDialog.0c9f9a568c',
-                'Supports Markdown formatting. Use Generate with AI to auto-fill from your changes.'
-              )}
-            </p>
-          </div>
-
-          <label className="flex items-center gap-2 rounded-md border border-border bg-background px-3 py-2 text-sm text-foreground transition-colors hover:bg-accent hover:text-accent-foreground">
-            <input
-              type="checkbox"
-              checked={draft}
-              onChange={(event) => setDraft(event.target.checked)}
-              className="size-4 shrink-0 rounded border-border accent-primary"
-            />
-            <span className="min-w-0 flex-1 truncate">
-              {translate(
-                'auto.components.right.sidebar.CreatePullRequestDialog.7ef56f3efe',
-                'Create as draft'
-              )}
-            </span>
-          </label>
-
-          {stripBaseRef(base).toLowerCase() === stripBaseRef(branch).toLowerCase() ? (
-            <p className="text-xs text-destructive">
-              {translate(
-                'auto.components.right.sidebar.CreatePullRequestDialog.27ef4b195c',
-                'Choose a different base branch before creating a'
-              )}
-              {copy.reviewLabel}.
-            </p>
-          ) : null}
-          {generateError ? <p className="text-xs text-destructive">{generateError}</p> : null}
-          {error ? <p className="text-xs text-destructive">{error}</p> : null}
-        </div>
+        <CreatePullRequestDialogForm
+          branch={branch}
+          base={base}
+          setBase={setBase}
+          baseQuery={baseQuery}
+          setBaseQuery={setBaseQuery}
+          baseResults={baseResults}
+          setBaseResults={setBaseResults}
+          baseSearchError={baseSearchError}
+          title={title}
+          setTitle={setTitle}
+          body={body}
+          setBody={setBody}
+          draft={draft}
+          setDraft={setDraft}
+          copy={copy}
+          generateError={generateError}
+          error={error}
+        />
 
         <DialogFooter>
           <Button variant="outline" onClick={() => handleOpenChange(false)} disabled={submitting}>
